Reorganized and removed dead code
