+++ src/components/Topbar.js
@@ -3,87 +3,33 @@ import { css } from '@emotion/react';
 import { SignInButton, useAuth, SignedIn, SignedOut, UserButton } from '@clerk/nextjs';
 export default function TopBar() {
     const { isLoaded, userId, sessionId, getToken } = useAuth();
+    function burgerActive() {
+        const navbarMenu = document.querySelector('#nav-links');
+        navbarMenu.classList.toggle('is-active');
+    }
     return (
-        <div
-            css={css`
-                position: fixed;
-                top: 0;
-                width: 100%;
-                height: 5rem;
-                display: flex;
-                flex-direction: row;
-                justify-content: space-evenly;
-                background-color: slategray;
-            `}>
-            <div
-                css={css`
-                    flex: 1;
-                    display: flex;
-                    justify-content: center;
-                `}>
-                <p
-                    css={css`
-                        margin: auto;
-                    `}>
-                    Logo
-                </p>
+        <nav className='navbar is-fixed-top is-dark' role='navigation' aria-label='main navigation'>
+            <div className='navbar-brand'>
+                <a className='navbar-item'>YOOOOOO PLANT APP</a>
+                <a className='navbar-burger' id='burger' onClick={burgerActive}>
+                    <span></span>
+                    <span></span>
+                    <span></span>
+                </a>
             </div>
-            <div
-                css={css`
-                    flex: 1;
-                    display: flex;
-                    justify-content: center;
-                `}>
-                <p
-                    css={css`
-                        margin: auto;
-                    `}>
-                    Search
-                </p>
+            <div className='navbar-menu ' id='nav-links'>
+                <div className='navbar-end '>
+                    <a className='navbar-item '>Settings</a>
+                    <a className='navbar-item '>
+                        <SignedIn>
+                            <UserButton></UserButton>
+                        </SignedIn>
+                        <SignedOut>
+                            <SignInButton></SignInButton>
+                        </SignedOut>
+                    </a>
+                </div>
             </div>
-            <div
-                css={css`
-                    flex: 1;
-                    display: flex;
-                    justify-content: center;
-                `}>
-                <p
-                    css={css`
-                        margin: auto;
-                    `}>
-                    Menu
-                </p>
-            </div>
-            <div
-                css={css`
-                    flex: 1;
-                    display: flex;
-                    justify-content: center;
-                `}>
-                <SignedIn>
-                    {/* Mount the UserButton component */}
-
-                    <UserButton />
-                </SignedIn>
-                <SignedOut>
-                    {/* Signed out users get sign in button */}
-
-                    <SignInButton />
-                </SignedOut>
-                {/* <SignedOut>
-                    <SignInButton mode='modal'>
-                        <button className='btn'>Sign up</button>
-                    </SignInButton>
-                </SignedOut>
-                <SignedIn>
-                    <p
-                        css={css`
-                            margin: auto;
-                        `}>
-                        Hello User
-                    </p>
-                </SignedIn> */}
-            </div>
-        </div>
-    );
+        </nav>
+    )
 }
