+++ src/components/layout.js
@@ -1,12 +1,13 @@
 //https://nextjs.org/docs/basic-features/layouts
-
-import Navbar from './Topbar';
+/** @jsxImportSource @emotion/react */
+import { css } from '@emotion/react';
+import Topbar from './Topbar';
 
 export default function Layout({ children }) {
     return (
         <>
-            <Navbar />
-            <main>{children}</main>
+            <Topbar />
+            <main css={css`margin-top: 5rem;`}>{children}</main>
         </>
     );
 }
